restructured components for optimized organization
